… core/GlobalState.h
… core/Loc.cc
… core/Loc.h
… core/Names.cc
… core/Symbols.cc
… core/TypePtr.cc
… core/TypePtr.h
… core/serialize/serialize.cc
… infer/environment.cc
… main/lsp/ErrorReporter.cc
… main/pipeline/pipeline.cc
… resolver/resolver.cc