<!-- (optional) Explain your change, focusing on the details of the solution. This is a great place to call out user-visible changes. -->


### Motivation
<!-- Why make this change? Describe the problem, not the solution. This can also be a link to an issue. -->

It had been starting to annoy me how long it would take to start Sorbet in debug
mode, so I went and profiled it and culled the biggest offenders.


### Test plan
<!-- If you did not write tests for this change, replace the message below explaining why not. Why we should be confident this change is correct? If you changed the website, please include a screenshot of the proposed changes. -->

Existing tests

```
~/sorbet (master)
❯ hyperfine --warmup 3 "bazel-bin/main/sorbet -e 0"
Benchmark 1: bazel-bin/main/sorbet -e 0
  Time (mean ± σ):     959.5 ms ±  13.4 ms    [User: 930.1 ms, System: 20.4 ms]
  Range (min … max):   939.5 ms … 985.7 ms    10 runs

~/sorbet (jez-timers)
❯ hyperfine --warmup 3 "bazel-bin/main/sorbet -e 0"
Benchmark 1: bazel-bin/main/sorbet -e 0
  Time (mean ± σ):     532.9 ms ±  15.0 ms    [User: 504.4 ms, System: 18.8 ms]
  Range (min … max):   520.3 ms … 567.2 ms    10 runs
```